Remove button and input elements, rename data-editing directory
