--- lib/components/button/button.module.scss
@@ -1,17 +1,0 @@
-.button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0.25rem 0.5rem;
-  border-radius: 0.25rem;
-  cursor: pointer;
-
-  &:hover:not([disabled]) {
-    background-image: linear-gradient(rgb(0 0 0 / 5%) 0 0);
-  }
-
-  &[disabled] {
-    filter: brightness(90%);
-    cursor: auto;
-  }
-}
