+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@portfolijo/react-comp-lib",
   "private": false,
-  "version": "1.2.0",
+  "version": "2.0.0",
   "type": "module",
   "main": "dist/main.js",
   "types": "dist/main.d.ts",
