+++ src/App.tsx
@@ -1,8 +1,6 @@
 import './App.css';
 import { SkipNavDemo } from './skip-nav-demo/skip-nav-demo';
 import { OutsideClickHandlerDemo } from './outside-click-handler-demo/outside-click-handler-demo';
-import { ButtonDemo } from './button-demo/button-demo';
-import { InputDemo } from './input-demo/input-demo';
 import { TableDemo } from './table-demo/table-demo';
 import { ExpandingContextDemo } from './expanding-context-demo/expanding-context-demo';
 import { DataEditingContextDemo } from './data-editing-context-demo/data-editing-context-demo';
@@ -13,8 +11,6 @@ function App() {
       {/* FIXME: SkipNavDemo must be first, or else it won't work. */}
       <SkipNavDemo containerId="demos" />
       <OutsideClickHandlerDemo />
-      <ButtonDemo />
-      <InputDemo />
       <TableDemo />
       <ExpandingContextDemo />
       <DataEditingContextDemo />
