--- src/input-demo/input-demo.module.scss
@@ -1,5 +1,0 @@
-.examples {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 1.5rem;
-}
